[FLINK-36369][table] Move deprecated user-visible classes in table-common to the legacy package
